requirements.txt
python-dotenv==0.10.3
python-telegram-bot==11.1.0
requests==2.22.0
